PlanningUI offers access to DrinkEditor and LayoutEditor
