+++ Assets/Scripts/UI/LayoutEditorUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 887e55d9ce9b04057a40a5b13d4b7397
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
